agent split by files
